# Build: Greatly improve performance of 'source and transform nodes' step 