--- a/.gitignore
+++ b/.gitignore
@@ -2,6 +2,13 @@
 .cache
 node_modules
 yarn-error.log
+.pnp.*
+.yarn/*
+!.yarn/patches
+!.yarn/plugins
+!.yarn/releases
+!.yarn/sdks
+!.yarn/versions
 
 # Build directory
 /public