--- a/plugins/gatsby-plugin-git-info/.gitignore
+++ b/plugins/gatsby-plugin-git-info/.gitignore
@@ -1,0 +1,9 @@
+# Project dependencies
+node_modules
+.pnp.*
+.yarn/*
+!.yarn/patches
+!.yarn/plugins
+!.yarn/releases
+!.yarn/sdks
+!.yarn/versions