## Issue

Hey PostHog team! Just started playing around and getting familiar with the website, but I noticed that during the build process the 'source and transform nodes' step always seemed to take an incredibly long time. I did some profiling and tracked the issue down to the `gatsby-transformer-gitinfo` plugin. This package was very poorly optimized and small in scope, so I went ahead and created a local `gatsby-plugin-git-info` plugin which does the same thing but in much less time. On my setup I went from this step taking regularly over **140s** to typically less than **35s**!

## Primary Change
- Create the `gatsby-plugin-git-info` package to replace `gatsby-transformer-gitinfo`
## Other small tweaks
- Ignoring yarn v2 specific files
- Adding the `plugins` directories as workspaces to make dependency management easier

This is my first PR here so please let me know if there's anything more you need or tweaks you want me to make. Super excited to be onboard!