@pjhul 2022-06-29
Forgot that in yarn classic you can't use worspaces in a non-private package. Very strange limitation but anyways. Even though this package is in a public github repo, setting it to private shouldn't make a difference since we don't publish it as a standalone package.