@pjhul 2022-07-01
Did a bit of testing comparing the `include` option in the gatsby-transformer-gitinfo and the fresh plugin I built. While adding this option drastically speeds up the transform step, I did still find the new version to be faster. On my setup running a completely fresh build these seemed like the average times. 

| Current | Current (with `include` option)      | New (including onPreInit step) |
| -------- | ----------- | ----------- |
| ~180s | ~65s      | ~45s      |